Separate space files
